Move toolchain tests to own file
